# back/docs/swagger.yaml
basePath: /api/v1
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Game:
    properties:
      created_at:
        type: string
      id:
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Match:
    properties:
      created_at:
        type: string
      endTime:
        type: string
      id:
        type: integer
      matchPosition:
        type: integer
      playerOne:
        $ref: '#/definitions/models.User'
      playerOneID:
        type: integer
      playerTwo:
        $ref: '#/definitions/models.User'
      playerTwoID:
        type: integer
      scores:
        items:
          $ref: '#/definitions/models.Score'
        type: array
      startTime:
        type: string
      status:
        type: string
      tournament:
        $ref: '#/definitions/models.Tournament'
      tournamentID:
        type: integer
      tournamentStep:
        $ref: '#/definitions/models.TournamentStep'
      tournamentStepID:
        type: integer
      updated_at:
        type: string
      winner:
        $ref: '#/definitions/models.User'
      winnerID:
        type: integer
    required:
    - status
    type: object
  models.Media:
    properties:
      created_at:
        type: string
      file_extension:
        type: string
      file_name:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Score:
    properties:
      created_at:
        type: string
      id:
        type: integer
      match:
        $ref: '#/definitions/models.Match'
      matchID:
        type: integer
      player:
        $ref: '#/definitions/models.User'
      playerID:
        type: integer
      score:
        type: integer
      updated_at:
        type: string
    type: object
  models.Tag:
    properties:
      created_at:
        type: string
      games:
        items:
          $ref: '#/definitions/models.Game'
        type: array
      id:
        type: integer
      name:
        type: string
      tournaments:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
  models.Tournament:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      game:
        $ref: '#/definitions/models.Game'
      game_id:
        type: integer
      id:
        type: integer
      location:
        type: string
      maxPlayers:
        example: 32
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      name:
        type: string
      organizer_id:
        type: integer
      rounds:
        type: integer
      start_date:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tournament_steps:
        items:
          $ref: '#/definitions/models.TournamentStep'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - description
    - end_date
    - game_id
    - maxPlayers
    - name
    - rounds
    - start_date
    type: object
  models.TournamentStep:
    properties:
      created_at:
        type: string
      id:
        type: integer
      matches:
        items:
          $ref: '#/definitions/models.Match'
        type: array
      name:
        type: string
      sequence:
        type: integer
      tournamentID:
        type: integer
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      address:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      matches:
        items:
          $ref: '#/definitions/models.Match'
        type: array
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        type: string
      tournaments:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      updated_at:
        type: string
      username:
        type: string
    required:
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /tournaments:
    get:
      consumes:
      - application/json
      description: Get all tournaments
      parameters:
      - description: Add recent tournaments to the response
        in: query
        name: WithRecents
        type: boolean
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tournaments
      tags:
      - tournament
    post:
      consumes:
      - multipart/form-data
      description: Create a new tournament
      parameters:
      - description: Tournament name
        example: s
        in: formData
        name: name
        required: true
        type: string
      - description: Tournament description
        example: s
        in: formData
        name: description
        required: true
        type: string
      - description: Tournament start date
        example: s
        in: formData
        name: start_date
        required: true
        type: string
      - description: Tournament end date
        example: s
        in: formData
        name: end_date
        required: true
        type: string
      - description: Organizer ID
        in: formData
        name: organizer_id
        required: true
        type: integer
      - description: Game ID
        in: formData
        name: game_id
        required: true
        type: integer
      - description: Number of rounds
        in: formData
        name: rounds
        required: true
        type: integer
      - collectionFormat: csv
        description: Array of tag IDs
        in: formData
        items:
          type: integer
        name: tagsIDs[]
        required: true
        type: array
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      - description: Location
        example: s
        in: formData
        name: location
        required: true
        type: string
      - description: Maximum number of players
        in: formData
        name: max_players
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tournament'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a new tournament
      tags:
      - tournament
  /tournaments/{id}:
    get:
      consumes:
      - application/json
      description: Get a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tournament'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a tournament
      tags:
      - tournament
  /tournaments/{id}/generate-matches:
    post:
      consumes:
      - application/json
      description: Generate matches for a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate matches for a tournament
      tags:
      - tournament
  /tournaments/{id}/matches:
    get:
      consumes:
      - application/json
      description: Get all matches of a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all matches of a tournament
      tags:
      - tournament
  /tournaments/{id}/register/{userID}:
    post:
      consumes:
      - application/json
      description: Register a user to a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a user to a tournament
      tags:
      - tournament
  /tournaments/{id}/start:
    post:
      consumes:
      - application/json
      description: Start a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth ini: []
      summary: Start a tournament
      tags:
      - tournament
  /tournaments/rankings:
    get:
      consumes:
      - application/json
      description: Get the ranking of a tournament
      parameters:
      - description: Tournament ID
        in: query
        name: TournamentID
        type: integer
      - description: Game ID
        in: query
        name: GameID
        type: integer
      - description: User ID
        in: query
        name: UserID
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the ranking of a tournament
      tags:
      - tournament
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"